pcml/functions/cbt_datagen/requirements.txt
# A reduced set of requirements enough to load and tf.Eager pcml models but not
# for full use of the library. In the interest of faster deployment (dealing with
# deployment timeout).
numpy==1.16.2
requests==2.21.0
scipy==1.2.1
six==1.12.0
tensor2tensor==1.14.0
tensorflow==1.14.0
google-auth
requests-toolbelt
sklearn
kubernetes
google-cloud-bigtable
Pillow==6.0.0
opencv-python==4.1.0.25
Flask==1.0.2
google-cloud-logging
gast==0.2.2